replace cutlass submodule references with explicit build step
this has the benefit of not weighing down CI builds on non-CUDA
platforms, and also enables shallow clone to make builds on CUDA
platforms lighter weight. The cutlass checkout alone is 210MiB,
which this commit saves on platforms that don't need it.

I don't have a modern enough CUDA GPU available to run with
flash-attention, but I have tested on older CUDA machine as well
as on MacOS, and modulo the asserts for the NVIDIA hardware version,
the build completes as expected.